Create 11.EqualNeightbours.js
A function that finds the number of equal neighbor pairs inside a matrix of variable size and type (numbers or strings).
The input comes as array of arrays, containing string elements (2D matrix of strings).
The output is return value of your function. Save the number of equal pairs you find and return it.