Create 06.JSONtoHTML.js
You’re tasked with creating an HTML table of students and their scores. You will receive a single string representing an array of objects, the table’s headings should be equal to the object’s keys, while each object’s values should be a new entry in the table. Any text values in an object should be escaped, in order to avoid introducing dangerous code into the HTML. 
Input
The input comes a single string argument (the array of objects).
Output
The output should be printed on the console – for each entry row in the input print the object representing it.