Create 01.CalorieObject.js
A function that composes an object by given properties. The input comes as an array of strings. Every even index of the array represents the name of the food. Every odd index is a number that is equal to the calories in 100 grams of the given product. Assign each value to its corresponding property and print it on the console.
The input comes as an array of string elements.
The output should be printed on the console.