Create 08.Rectangle.js
Write a function that creats and returns a rectangle object. The rectangle needs to have a width (Number), height (Number) and color (String) properties, which are set via arguments during creation, and a calcArea() method, that calculates and returns the rectangle’s area.
Input
The function will receive three valid parameters – width (Number), height (Number) and color (String).
Output
Your function must return an object with all properties and methods as described. The calcArea() method of the object should return a number.